requirements.txt
Flask
github3.py
pandas
flask-cors
